% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_example_project.R
\name{download_example_project}
\alias{download_example_project}
\title{Download Example Project}
\usage{
download_example_project(.dir, .name)
}
\arguments{
\item{.dir}{Full Path to the Directory}

\item{.name}{Name of the Project}
}
\value{
A Folder containing Example Files
}
\description{
Download Example Project
}
